% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deftly.R
\name{deftly_demo}
\alias{deftly_demo}
\title{Create a data.frame in C using the \code{deft} package}
\usage{
deftly_demo(n = 8)
}
\arguments{
\item{n}{number of rows to return in the demo data.frame}
}
\value{
data.frame created in C and returned to R
}
\description{
Create a data.frame in C using the \code{deft} package
}
